--- a/src/utilities/include/vec_dh.h
+++ b/src/utilities/include/vec_dh.h
@@ -401,12 +401,10 @@ class VecDH {
   const T *ptrH() const { return cptrH(); }
 
   T &operator[](int i) {
-    impl_.prefetch_to(true);
     return impl_[i];
   }
 
   const T &operator[](int i) const {
-    impl_.prefetch_to(true);
     return impl_[i];
   }