--- a/src/utilities/include/vec_dh.h
+++ b/src/utilities/include/vec_dh.h
@@ -400,13 +400,9 @@ class VecDH {
 
   const T *ptrH() const { return cptrH(); }
 
-  T &operator[](int i) {
-    return impl_[i];
-  }
+  T &operator[](int i) { return impl_[i]; }
 
-  const T &operator[](int i) const {
-    return impl_[i];
-  }
+  const T &operator[](int i) const { return impl_[i]; }
 
   T &back() { return impl_.back(); }