@pca006132 2023-07-15
I haven't check the performance with CUDA because for some reason it cannot detect CUDA on my machine. Not sure if this is my environment problem or bug in our code.