@pca006132 2023-07-15
I think we don't have to care about CUDA performance for now --- we already know that it is not very good. We can probably revisit it later when we finally want to care about CUDA.